Check login and admin

+++ SWP391.E.BL5-G1/src/java/admin/DeleteAbout.java
@@ -10,6 +10,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import java.io.IOException;
 import java.io.PrintWriter;
 
@@ -58,6 +59,17 @@ public class DeleteAbout extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        HttpSession session = request.getSession(false);
+        if (session == null || session.getAttribute("user") == null) {
+            response.sendRedirect("user?action=login");
+            return;
+        }
+
+        model.User user = (model.User) session.getAttribute("user");
+        if (!user.getIsStoreStaff().equals("1")) {
+            response.sendRedirect("home");
+            return;
+        }
         String aboutId = request.getParameter("id");
         aboutDAO dao = new aboutDAO();
         dao.deleteAbout(aboutId);
